change some datatype and organize format

+++ database/migrations/2021_06_21_060039_create_apptables_table.php
@@ -25,7 +25,7 @@ class CreateApptablesTable extends Migration
             $table->unsignedBigInteger('season_id')->nullable();
             $table->foreign('season_id')->references('season_id')->on('seasons')->onUpdate('cascade')->onDelete('set null');
             $table->string('name', 50);
-            $table->integer('amount');
+            $table->double('amount');
             $table->timestamp('date')->useCurrent();
         });
 
@@ -34,8 +34,8 @@ class CreateApptablesTable extends Migration
             $table->unsignedBigInteger('season_id')->nullable();
             $table->foreign('season_id')->references('season_id')->on('seasons')->onUpdate('cascade')->onDelete('set null');
             $table->string('name', 50);
-            $table->integer('quantity');
-            $table->integer('cost');
+            $table->double('quantity');
+            $table->double('cost');
             $table->timestamp('date')->useCurrent();
         });
 
@@ -51,7 +51,7 @@ class CreateApptablesTable extends Migration
             $table->foreign('season_id')->references('season_id')->on('seasons')->onUpdate('cascade')->onDelete('set null');
             $table->unsignedBigInteger('laborer_id')->nullable();
             $table->foreign('laborer_id')->references('laborer_id')->on('laborers')->onUpdate('cascade')->onDelete('set null');
-            $table->integer('wage');
+            $table->double('wage');
             $table->timestamp('date')->useCurrent();
         });
 
@@ -59,9 +59,9 @@ class CreateApptablesTable extends Migration
             $table->id('revenue_id');
             $table->unsignedBigInteger('season_id')->nullable();
             $table->foreign('season_id')->references('season_id')->on('seasons')->onUpdate('cascade')->onDelete('set null');
-            $table->integer('quantity');
-            $table->integer('price_per_unit');
-            $table->integer('total_price');
+            $table->double('quantity');
+            $table->double('price_per_unit');
+            $table->double('total_price');
             $table->timestamp('date')->useCurrent();
         });
 
@@ -69,7 +69,7 @@ class CreateApptablesTable extends Migration
             $table->id('yield_id');
             $table->unsignedBigInteger('season_id')->nullable();
             $table->foreign('season_id')->references('season_id')->on('seasons')->onUpdate('cascade')->onDelete('set null');
-            $table->integer('quantity');
+            $table->double('quantity');
             $table->timestamp('date')->useCurrent();
         });
